[CI][AMD] Consolidate generator interface
